feat: Extracting change log generation with its request
